Separate PassageProcessor
